Plumbed docid_base_addr through merge iterator, etc.

+++ imhotep-server/src/main/native/optimized_ftgs/optimized/include/split_view.hpp
@@ -63,7 +63,7 @@ namespace imhotep {
         const size_t            id_size(read<size_t>());
         const SplitView::Buffer id(read_bytes(id_size));
         const int64_t           doc_offset(read<int64_t>());
-        const int64_t           doc_freq(read<int64_t>());
+        const int32_t           doc_freq(read<int32_t>());
         return StringTerm(std::string(id.first, id.second), doc_offset, doc_freq);
     }
 }
